Split client.py and server.py main functions into separate files
